+++ src/components/analytics/DashStats.tsx
@@ -1,7 +1,8 @@
-import React, { Suspense } from 'react';
-import { createStyles, Loader, Select, Text } from '@mantine/core';
+import React, { Dispatch, SetStateAction, Suspense } from 'react';
+import { Button, createStyles, Group, Loader, Select, Text, Title, useMantineTheme } from '@mantine/core';
 import { getMetrics, Metrics } from '@/services/analytics';
 import { useQuery } from '@tanstack/react-query';
+import { queryClient } from '@/services/queryClient';
 
 const useStyles = createStyles((theme) => ({
   root: {
@@ -59,19 +60,48 @@ const useStyles = createStyles((theme) => ({
   },
 }));
 
-type Props = {
-site: string,
-period: string}
 
-export function DashStats({site, period}: Props ) {
+export function DashStats() {
+  const [site, setSite] = React.useState<string | null>('jmwalsh.dev');
+  const [period, setPeriod] = React.useState<string | null>('day');
   const { classes } = useStyles();
+  const theme = useMantineTheme();
   const { data, error } = useQuery<Metrics, Error>(["metrics", site, period], () =>
-    getMetrics(site, period)
-  );
-
-/* function handleSiteChange(e: React.ChangeEvent<HTMLSelectElement>) { */
+  getMetrics(site, period)
+);
 
   return (
+    <>
+    <Group position="apart" mb={"lg"} sx={{
+        padding: "1rem",
+        border: `3px solid ${theme.colors.primary[3]}`
+      }}>
+        <Title>{site}</Title>
+      <Select
+      label="Site"
+      value={site}
+      width={200}
+      data={[
+        { value: 'jmwalsh.dev', label: 'Portfolio' },
+        { value: 'cryptones.vercel.app', label: 'crypTones' },
+      ]}
+      onChange={setSite}
+      />
+      <Select
+      label="Time period"
+      value={period}
+      width={200}
+      data={[
+        { value: 'day', label: 'Today' },
+        { value: '7d', label: '7 days' },
+        { value: '30d', label: '30 days' },
+        { value: 'month', label: 'Current Month' },
+      ]}
+      onChange={setPeriod}
+      />
+   
+      </Group>
+
     <div className={classes.root}>
 
 
@@ -106,5 +136,6 @@ export function DashStats({site, period}: Props ) {
 {/*          <Text className={classes.description}></Text> */}
        </div>
     </div>
+    </>
 )
 }
